Moved the Token definition into `types.rs`
